+++ src/app/routes/payment/movieTicket.ts
@@ -49,11 +49,28 @@ movieTicketPaymentRouter.post(
                 throw new cinerino.factory.errors.ServiceUnavailable('Project settings not found');
             }
 
+            let paymentMethodType: cinerino.factory.paymentMethodType.MovieTicket | cinerino.factory.paymentMethodType.MGTicket
+                = req.body.typeOf;
+            if (typeof paymentMethodType !== 'string') {
+                paymentMethodType = cinerino.factory.paymentMethodType.MovieTicket;
+            }
+
             const action = await cinerino.service.payment.movieTicket.checkMovieTicket({
                 project: req.project,
                 typeOf: cinerino.factory.actionType.CheckAction,
                 agent: req.agent,
-                object: req.body
+                object: {
+                    ...req.body,
+                    movieTickets: (Array.isArray(req.body.movieTickets))
+                        ? (<any[]>req.body.movieTickets).map((m) => {
+                            return {
+                                ...m,
+                                typeOf: paymentMethodType
+                            };
+                        })
+                        : [],
+                    typeOf: paymentMethodType
+                }
             })({
                 action: new cinerino.repository.Action(mongoose.connection),
                 project: new cinerino.repository.Project(mongoose.connection),
@@ -138,17 +155,28 @@ movieTicketPaymentRouter.post<ParamsDictionary>(
                 throw new cinerino.factory.errors.ServiceUnavailable('Project settings not found');
             }
 
+            let paymentMethodType: cinerino.factory.paymentMethodType.MovieTicket | cinerino.factory.paymentMethodType.MGTicket
+                = req.body.object?.typeOf;
+            if (typeof paymentMethodType !== 'string') {
+                paymentMethodType = cinerino.factory.paymentMethodType.MovieTicket;
+            }
+
             const action = await cinerino.service.payment.movieTicket.authorize({
                 agent: { id: req.user.sub },
                 object: {
-                    typeOf: cinerino.factory.paymentMethodType.MovieTicket,
+                    typeOf: <any>paymentMethodType,
                     amount: 0,
                     additionalProperty: (Array.isArray(req.body.object.additionalProperty))
                         ? (<any[]>req.body.object.additionalProperty).map((p: any) => {
                             return { name: String(p.name), value: String(p.value) };
                         })
                         : [],
-                    movieTickets: req.body.object.movieTickets,
+                    movieTickets: (<any[]>req.body.object.movieTickets).map((o) => {
+                        return {
+                            ...o,
+                            typeOf: paymentMethodType
+                        };
+                    }),
                     ...(typeof req.body.object.name === 'string') ? { name: <string>req.body.object.name } : undefined
                 },
                 purpose: { typeOf: req.body.purpose.typeOf, id: <string>req.body.purpose.id }