+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - update @cinerino/domain
 - プロジェクトのuseReservationNumberAsConfirmationNumber設定を削除
+- オファーと価格仕様のappliesToMovieTicketType→appliesToMovieTicket対応
 
 ### Deprecated