+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update @cinerino/domain
 - プロジェクトのuseReservationNumberAsConfirmationNumber設定を削除
 - オファーと価格仕様のappliesToMovieTicketType→appliesToMovieTicket対応
+- ムビチケ系統決済処理をMGTicketに対応
 
 ### Deprecated